redis cache

+++ controllers/ticketController.js
@@ -1,27 +1,35 @@
 const supabase = require('../supabaseClient.js')
+const redisCaching = require('../redisCaching.js');
 
 async function addTicket(req, res) {
     const ticket = req.body
     const response = await supabase.from('ticket').insert(ticket).select()
+
+    redisCaching.removeData(`tickets:${req.body.creator}`)
+
     res.status(200).json(response)
 }
 
 async function deleteTicket(req, res) {
     const ticketID = req.body.ticketID
 
-    const response = await supabase.from('ticket').delete().eq('id', ticketID)
+    const response = await supabase.from('ticket').delete().eq('id', ticketID).select('creator')
+    redisCaching.removeData(`tickets:${response.data[0].creator}`)
 
     res.status(200).json(response)
 }
 
 async function getUserTicket(req, res) {
 
-    const response = await supabase
-                            .from('ticket')
-                            .select()
-                            .eq('creator', req.query.user_id)
+    const tickets = await redisCaching.getOrSetCache(`tickets:${req.query.user_id}`, async () => {
 
-    res.status(200).json(response)
+        return await supabase
+            .from('ticket')
+            .select()
+            .eq('creator', req.query.user_id)
+    })
+
+    res.status(200).json(tickets)
 }
 
 
@@ -41,7 +49,14 @@ async function changeStatus(req, res) {
     const newStatus = req.body.newStatus
     const admin_comment = req.body.admin_comment
 
-    const {data, error} = await supabase.from('ticket').update({ status: newStatus, admin_comment }).eq('id', ticketID).select();
+    const response = await supabase
+        .from('ticket')
+        .update({ status: newStatus })
+        .eq('id', ticketID)
+        .select('creator')
+
+    redisCaching.removeData(`tickets:${response.data[0].creator}`)
+
 
     if (error) res.json(error)
     else res.json(data)
